+++ kool-core/src/commonMain/kotlin/de/fabmax/kool/modules/mesh/simplification/TermCriterion.kt
@@ -10,19 +10,19 @@ import kotlin.math.round
  */
 interface TermCriterion {
     fun init(mesh: HalfEdgeMesh) { }
-    fun isFinished(mesh: HalfEdgeMesh, nextError: Float): Boolean
+    fun isFinished(mesh: HalfEdgeMesh, nextError: Double): Boolean
 }
 
-fun terminateOnFaceCountRel(factor: Float) = object : TermCriterion {
+fun terminateOnFaceCountRel(factor: Double) = object : TermCriterion {
     var targetFaceCnt = Int.MAX_VALUE
     override fun init(mesh: HalfEdgeMesh) { targetFaceCnt = round(mesh.faceCount * factor).toInt() }
-    override fun isFinished(mesh: HalfEdgeMesh, nextError: Float) = mesh.faceCount <= targetFaceCnt
+    override fun isFinished(mesh: HalfEdgeMesh, nextError: Double) = mesh.faceCount <= targetFaceCnt
 }
 
 fun terminateOnFaceCountAbs(targetFaceCnt: Int) = object : TermCriterion {
-    override fun isFinished(mesh: HalfEdgeMesh, nextError: Float) = mesh.faceCount <= targetFaceCnt
+    override fun isFinished(mesh: HalfEdgeMesh, nextError: Double) = mesh.faceCount <= targetFaceCnt
 }
 
 fun terminateOnError(targetError: Double) = object : TermCriterion {
-    override fun isFinished(mesh: HalfEdgeMesh, nextError: Float) = nextError > targetError
+    override fun isFinished(mesh: HalfEdgeMesh, nextError: Double) = nextError > targetError
 }
